+++ TvEngine3/TVLibrary/TVLibrary/Interfaces/Analyzer/ITsEpgScanner.cs
@@ -127,7 +127,7 @@ namespace TvLibrary.Interfaces.Analyzer
     /// <param name="count">The count.</param>
     /// <returns></returns>
     [PreserveSig]
-    int GetMHWTitleCount(out Int16 count);
+    int GetMHWTitleCount(out uint count);
 
     /// <summary>
     /// Gets the details for a MHW title.
@@ -148,7 +148,7 @@ namespace TvLibrary.Interfaces.Analyzer
     /// <param name="programName">Name of the program.</param>
     /// <returns></returns>
     [PreserveSig]
-    int GetMHWTitle(UInt16 program, ref uint id, ref UInt16 transportId, ref UInt16 networkId, ref UInt16 channelId, ref uint programId, ref UInt16 themeId, ref UInt16 PPV, ref byte Summaries, ref UInt16 duration, ref uint dateStart, ref uint timeStart, out IntPtr title, out IntPtr programName);
+    int GetMHWTitle(uint program, ref uint id, ref uint transportId, ref uint networkId, ref uint channelId, ref UInt32 programId, ref uint themeId, ref uint PPV, ref byte Summaries, ref uint duration, ref uint dateStart, ref uint timeStart, out IntPtr title, out IntPtr programName);
 
     /// <summary>
     /// Gets the details for a MHW channel.
@@ -160,7 +160,7 @@ namespace TvLibrary.Interfaces.Analyzer
     /// <param name="channelName">Name of the channel.</param>
     /// <returns></returns>
     [PreserveSig]
-    int GetMHWChannel(UInt16 channelNr, ref UInt16 channelId, ref UInt16 networkId, ref UInt16 transportId, out IntPtr channelName);
+    int GetMHWChannel(uint channelNr, ref uint channelId, ref uint networkId, ref uint transportId, out IntPtr channelName);
 
     /// <summary>
     /// Gets the MHW summary.
@@ -169,7 +169,7 @@ namespace TvLibrary.Interfaces.Analyzer
     /// <param name="summary">The summary.</param>
     /// <returns></returns>
     [PreserveSig]
-    int GetMHWSummary(uint programId, out IntPtr summary);
+    int GetMHWSummary(UInt32 programId, out IntPtr summary);
 
     /// <summary>
     /// Gets the MHW theme.
@@ -178,7 +178,7 @@ namespace TvLibrary.Interfaces.Analyzer
     /// <param name="theme">The theme.</param>
     /// <returns></returns>
     [PreserveSig]
-    int GetMHWTheme(UInt16 themeId, out IntPtr theme);
+    int GetMHWTheme(uint themeId, out IntPtr theme);
 
     /// <summary>
     /// Resets this MHW grabber.
